Quote omnibenchmark[s3] in install command
Bash and zsh interprets "omnibenchmark[s3]" as a file glob, instead of a string
in the command "pip install omnibenchmark[s3]".
This happens to work in Bash for bad reasons, but it does not work in zsh.

Instead, quote the package name, which is more correct and works on zsh.